返回浮点数

+++ application/common/service/Withdraw.php
@@ -22,6 +22,6 @@ class Withdraw extends CareyShop
      */
     public function getWithdrawFee()
     {
-        return ['withdraw_fee' => Config::get('withdraw_fee.value', 'system_info')];
+        return ['withdraw_fee' => (float)Config::get('withdraw_fee.value', 'system_info')];
     }
 }
